test: first test for command line parsing
initially this failed because the defaults are not transferring
over to serde_json but they are transferring over via Config.

the error was for the two different paths to building a default value of
ExchangeRateConfig, clap used the (60000, 5) values, serde didn't. Serde
didn't because it called the derived Default implementation.

included comments that the two default values should be kept in sync.
luckily tests should start failing if only one is changed.

Signed-off-by: Joonas Koivunen <joonas.koivunen@gmail.com>